meh. celcius is now displayed as an int

+++ FWeather/plugin.py
@@ -86,8 +86,8 @@ class FWeather(callbacks.Plugin):
             try:
                 w = thefuckingweather.get_weather(text)
                 location = w['location']
-                f_temp = self._color(float(w['current']['temperature']), 'f')
-                c_temp = self._color(round((5.0/9.0)*(w['current']['temperature']-32), 1), 'c')
+                f_temp = self._color(w['current']['temperature'], 'f')
+                c_temp = self._color(int(round((5.0/9.0)*(w['current']['temperature']-32), 1)), 'c')
                 weather = w['current']['weather'][0]
                 remark = w['current']['remark']
                 comment = "{0}: {1}°F, {2}°C?! {3}! ({4})".format(location, f_temp, c_temp, weather, remark)
